requirements.txt
python-telegram-bot==21.5
openai-whisper
ffmpeg-python
pydub
python-dotenv

